Added error messages to log_workout page. Moved models into routes.py
I had to move the models into routes.py to get this all to work. The error messaging is now working and I moved it into the log_workout template where I think it makes more sense from a UX perspective. Also, when I pass variables into the templates for each elif, do I need to pass in the empty string? I tried it without and it seems to work but not sure what correct practice is.

The next step is figuring out how to save and render values for an incomplete log_workout submission. I have been googling around for how to do that for dropdowns and radio buttons but nothing much of substance yet.